requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
dataclasses==0.8
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.43.0
django-cleanup==5.1.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
djongo==1.3.3
idna==2.10
kafka-python==2.0.2
numpy==1.19.3
oauthlib==3.1.0
pandas==1.1.4
py4j==0.10.7
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.2
pyspark==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.2.4
tqdm==4.51.0
urllib3==1.26.2
uWSGI==2.0.19.1
